Support more CSIs and give a better example in README
Support now italics and underline (and selective disabling thereof via
VT220), include the two also in CSS. Support for \e[90m to \e[97m
and \e[100m to \e[107m ("increased intensity foreground/background") as
well. Also give an example in the README that works out-of-the-box and
clarify that ancient versions of screen, present on today's generation
of undying iPods, do not support giving a file name to the "-L"
parameter.